M jsf-ri/src/com/sun/faces/application/Resource/ResourceImpl.java   - Fixed P1 findbugs issue: result of integer multiplication cast to long   - general cleanup
 M jsf-ri/src/com/sun/faces/application/Resource/ResourceCache.java
  - Fixed P1 findbugs issue: uncallable method defined within anonymous class

 M jsf-ri/src/com/sun/faces/application/resource/ResourceHandlerImpl.java
  - minor cleanup to libraryExists() method

 M jsf-ri/src/com/sun/faces/facelets/impl/DefaultFaceletContext.java
  - Fixed P2 findbugs issue: method invokes inefficient Number constructor
  - Fixed P1 findbugs issue: call to equals() comparing different types
   + the TemplateManager.equals() was only invoked within DefaultFaceletContext.
      Added managesClient and managesFacelet to query the specific details
      without having to overload equals().  
  - Generified DefaultFaceletContext
  - replaced usages of StringBuffer with StringBuilder
  - other minor cleanup

git-svn-id: https://svn.java.net/svn/mojarra~svn/trunk@5365 761efcf2-d34d-6b61-9145-99dcacc3edf1

+++ jsf-ri/src/com/sun/faces/application/resource/ResourceCache.java
@@ -139,12 +139,12 @@ public class ResourceCache {
         ServletContext sc = config.getServletContext();
         contextName = getServletContextIdentifier(sc);
         shutdown = false;
-        int checkPeriod = getCheckPeriod(config);
+        long checkPeriod = getCheckPeriod(config);
         if (checkPeriod >= 0) {
             resourceCache = new MultiKeyConcurrentHashMap<String,ResourceInfo>(30);
         }
         if (checkPeriod >= 1) {
-            initExecutor((checkPeriod * 1000 * 60));
+            initExecutor((checkPeriod * 1000L * 60L));
             initMonitors(sc);
         }
         ApplicationAssociate associate = ApplicationAssociate.getInstance(sc);
@@ -287,13 +287,13 @@ public class ResourceCache {
     }
 
 
-    private int getCheckPeriod(WebConfiguration webConfig) {
+    private Long getCheckPeriod(WebConfiguration webConfig) {
 
         String val = webConfig.getOptionValue(WebContextInitParameter.ResourceUpdateCheckPeriod);
         try {
-            return Integer.parseInt(val);
+            return Long.parseLong(val);
         } catch (NumberFormatException nfe) {
-            return Integer.parseInt(WebContextInitParameter.ResourceUpdateCheckPeriod.getDefaultValue());
+            return Long.parseLong(WebContextInitParameter.ResourceUpdateCheckPeriod.getDefaultValue());
         }
 
     }
